requirements.txt
streamlit
ultralytics
opencv-python-headless
pandas
matplotlib
numpy
Pillow